Changes to be committed:
	new file:   .babelrc
	modified:   package.json
	deleted:    src/dimension.js
	new file:   src/dimension/Dimension.js
	new file:   src/dimension/index.js
	deleted:    src/index.js
	deleted:    src/measure.js
	new file:   src/measure/Measure.js
	new file:   src/measure/index.js
	new file:   src/prefix/Prefix.js
	renamed:    src/prefix.js -> src/prefix/index.js
	new file:   src/unit/Unit.js
	renamed:    src/unit.js -> src/unit/index.js
	new file:   test/measure/Measure.js
	modified:   webpack.config.js
