+++ src/dimension/Dimension.js
@@ -1,0 +1,12 @@
+export class Dimension extends Array {
+
+    constructor(symbol, ...unit) {
+        super(...unit);
+        this.symbol = symbol;
+    }
+
+    toString() {
+        return this.symbol;
+    }
+
+}
