+++ webpack.config.js
@@ -1,14 +1,14 @@
 /* eslint-env node */
 const path = require('path');
-const {name, module: entry} = require('./package.json');
+const {name: library, module: entry} = require('./package.json');
 
 module.exports = {
     mode: 'production',
     entry: entry,
     output: {
-        library: name,
+        library: library,
         libraryTarget: 'umd',
-        filename: `${name}.js`,
+        filename: `${library}.js`,
         path: path.resolve(__dirname, 'dist'),
     },
 };
